+++ draft-js-inline-toolbar-plugin/src/components/Separator/index.js
@@ -1,0 +1,4 @@
+import React from 'react';
+import styles from '../../inlineToolbarStyles.css';
+
+export default () => <div className={styles.separator} />;