+++ draft-js-plugins-editor/src/Editor/index.js
@@ -6,6 +6,7 @@ import {
   DefaultDraftBlockRenderMap,
 } from 'draft-js';
 import Immutable, { List, Map } from 'immutable';
+import MultiDecorator from './MultiDecorator';
 import createCompositeDecorator from './createCompositeDecorator';
 import moveSelectionToEnd from './moveSelectionToEnd';
 import proxies from './proxies';
@@ -51,6 +52,7 @@ class PluginEditor extends Component {
     this.initPlugins();
     this.decorateEditorState(this.props.editorState);
   }
+
 
   componentWillReceiveProps(nextProps) {
     const decorator = nextProps.editorState.getDecorator();
@@ -98,6 +100,8 @@ class PluginEditor extends Component {
     if (readOnly !== this.state.readOnly) this.setState({ readOnly });
   };
 
+  getEditorRef = () => this.editor;
+
   getEditorState = () => this.props.editorState;
   getPluginMethods = () => ({
     getPlugins: this.getPlugins,
@@ -106,6 +110,7 @@ class PluginEditor extends Component {
     getEditorState: this.getEditorState,
     getReadOnly: this.getReadOnly,
     setReadOnly: this.setReadOnly,
+    getEditorRef: this.getEditorRef,
   });
 
   initPlugins = () => {
@@ -129,12 +134,22 @@ class PluginEditor extends Component {
   };
 
   decorateEditorState = (editorState) => {
+    const decorators = this.resolveDecorators();
     const compositeDecorator = createCompositeDecorator(
-      this.resolveDecorators(),
+      decorators.filter((decorator) => !this.decoratorIsCustom(decorator)),
       this.getEditorState,
       this.onChange);
-    const decoratedEditorState = EditorState.set(editorState, { decorator: compositeDecorator });
-    this.onChange(moveSelectionToEnd(decoratedEditorState));
+
+    const customDecorators = decorators
+              .filter((decorator) => this.decoratorIsCustom(decorator));
+
+    const multiDecorator = new MultiDecorator([
+        ...customDecorators,
+        compositeDecorator,
+    ]);
+
+    const editorState = EditorState.set(this.props.editorState, { decorator: multiDecorator });
+    this.onChange(moveSelectionToEnd(editorState));
   };
 
   createEventHooks = (methodName, plugins) => (...args) => {
@@ -147,6 +162,18 @@ class PluginEditor extends Component {
     }
 
     return false;
+  };
+
+  createHandleHooks = (methodName, plugins) => (...args) => {
+    const newArgs = [].slice.apply(args);
+    newArgs.push(this.getPluginMethods());
+    for (const plugin of plugins) {
+      if (typeof plugin[methodName] !== 'function') continue;
+      const result = plugin[methodName](...newArgs);
+      if (result === 'handled') return 'handled';
+    }
+
+    return 'not-handled';
   };
 
   createFnHooks = (methodName, plugins) => (...args) => {
@@ -194,6 +221,7 @@ class PluginEditor extends Component {
   createPluginHooks = () => {
     const pluginHooks = {};
     const eventHookKeys = [];
+    const handleHookKeys = [];
     const fnHookKeys = [];
     const plugins = [this.props, ...this.resolvePlugins()];
 
@@ -204,9 +232,15 @@ class PluginEditor extends Component {
         // if `attrName` has been added as a hook key already, ignore this one
         if (eventHookKeys.indexOf(attrName) !== -1 || fnHookKeys.indexOf(attrName) !== -1) return;
 
-        const isEventHookKey = attrName.indexOf('on') === 0 || attrName.indexOf('handle') === 0;
+        const isEventHookKey = attrName.indexOf('on') === 0;
         if (isEventHookKey) {
           eventHookKeys.push(attrName);
+          return;
+        }
+
+        const isHandleHookKey = attrName.indexOf('handle') === 0;
+        if (isHandleHookKey) {
+          handleHookKeys.push(attrName);
           return;
         }
 
@@ -220,6 +254,10 @@ class PluginEditor extends Component {
 
     eventHookKeys.forEach((attrName) => {
       pluginHooks[attrName] = this.createEventHooks(attrName, plugins);
+    });
+
+    handleHookKeys.forEach((attrName) => {
+      pluginHooks[attrName] = this.createHandleHooks(attrName, plugins);
     });
 
     fnHookKeys.forEach((attrName) => {
@@ -245,6 +283,13 @@ class PluginEditor extends Component {
       .filter((plugin) => plugin.decorators !== undefined)
       .flatMap((plugin) => plugin.decorators);
   };
+
+  // Return true if decorator implements the DraftDecoratorType interface
+  // @see https://github.com/facebook/draft-js/blob/master/src/model/decorators/DraftDecoratorType.js
+  decoratorIsCustom = (decorator) => typeof decorator.getDecorations === 'function' &&
+    typeof decorator.getComponentForKey === 'function' &&
+    typeof decorator.getPropsForKey === 'function';
+
 
   resolveCustomStyleMap = () => (
     this.resolvePlugins()